refactor: 페이지 프로젝트 구조 변경
