--- src/components/Button/Button.style.tsx
@@ -1,9 +1,0 @@
-export const buttonColors = {
-  default:
-    'border-gray-500 text-gray-500 dark:text-white/80 hover:bg-gray-100 dark:hover:bg-gray-500/20',
-  blue: 'border-blue-500 text-blue-500 hover:bg-blue-100 dark:hover:bg-blue-500/20',
-  black:
-    'border-black bg-black text-white hover:bg-gray-900 dark:text-white/80 dark:hover:bg-gray-900',
-  red: 'border-red-500 bg-red-500 text-white hover:bg-red-900 dark:text-white/80 dark:hover:bg-red-900',
-  disabled: 'border bg-gray-200 text-white cursor-not-allowed',
-};
