+++ src/components/CaseElement.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren, useCallback, useEffect, useState } from 'react';
-import ElementLine from '../ElementLine/ElementLine';
-import Element from '../Element/Element';
+import ElementLine from './ElementLine';
+import Element from './Element';
 import { ElementType } from '@type/element';
 import { useSetCaseStore } from '@store/case';
 import { useSelectedCaseStore } from '@store/selected-case';
