--- src/components/Element/Element.style.tsx
@@ -1,5 +1,0 @@
-export const elementColors = {
-  layer: 'bg-sky-200 dark:bg-sky-200/80',
-  group: 'bg-gray-200 dark:bg-gray-200/80',
-  case: 'bg-white border-[1.5px] border-gray-300 dark:bg-white/80',
-};
