--- src/components/Element/Element.type.ts
@@ -1,1 +1,0 @@
-export type ElementColorType = 'layer' | 'group' | 'case';
