+++ src/pages/SplashPage/SplashPage.tsx
